+++ hydromt/_utils/__init__.py
@@ -1,5 +1,7 @@
 """Utility functions for hydromt that have no other home."""
 
+from osgeo import gdal
+
 from .caching import _cache_vrt_tiles, _copy_to_local
 from .dataset import (
     _rename_vars,
@@ -43,6 +45,8 @@ __all__ = [
     "_make_config_paths_abs",
     "_make_config_paths_relative",
 ]
+
+gdal.UseExceptions()
 
 
 class _classproperty(property):