+++ hydromt/_utils/caching.py
@@ -12,6 +12,7 @@ import geopandas as gpd
 import numpy as np
 from affine import Affine
 from fsspec import AbstractFileSystem, url_to_fs
+from osgeo import gdal
 from pyproj import CRS
 
 from hydromt._typing.type_def import StrPath
@@ -54,12 +55,12 @@ def _overlaps(source: ET.Element, affine: Affine, bbox: List[float]) -> bool:
     affine : Affine
         Tile affine
     bbox : List[float]
-        requested bbox
+        Requested bbox
 
     Returns
     -------
     bool
-        whether the tile overlaps with the requested bbox.
+        Whether the tile overlaps with the requested bbox.
 
     Raises
     ------
@@ -87,7 +88,7 @@ def _cache_vrt_tiles(
     fs: Optional[AbstractFileSystem] = None,
     geom: Optional[gpd.GeoSeries] = None,
     cache_dir: StrPath = SETTINGS.cache_root,
-) -> str:
+) -> Path:
     """Cache vrt tiles that intersect with geom.
 
     Note that the vrt file must contain relative tile paths.
@@ -95,18 +96,18 @@ def _cache_vrt_tiles(
     Parameters
     ----------
     vrt_uri: str
-        path to source vrt
+        Path to source vrt
     fs : Optional[AbstractFileSystem], optional
         Fsspec filesystem. Will be inferred from src if not supplied.
     geom: geopandas.GeoSeries, optional
-        geometry to intersect tiles with
+        Geometry to intersect tiles with
     cache_dir: str, Path
-        path of the root folder where
+        Path of the root folder where
 
     Returns
     -------
-    vrt_destination_path : str
-        path to cached vrt
+    vrt_destination_path : Path
+        Path to cached vrt
     """
     if fs is None:
         fs: AbstractFileSystem = url_to_fs(vrt_uri)[0]
@@ -117,12 +118,8 @@ def _cache_vrt_tiles(
     # Strip scheme from base uri and get path to directory with vrt file
     scheme, stripped = _strip_scheme(vrt_uri)
 
-    vrt_destination_path = join(cache_dir, basename(vrt_uri))
-    if not isfile(vrt_destination_path):
-        _copy_to_local(vrt_uri, vrt_destination_path, fs)
     # read vrt file
-    with open(vrt_destination_path, "r") as f:
-        root = ET.fromstring(f.read())
+    root = ET.fromstring(fs.read_text(vrt_uri))
 
     # get vrt transform and crs
     geotransform_el: Optional[ET.Element] = root.find("GeoTransform")
@@ -199,10 +196,44 @@ def _cache_vrt_tiles(
             if not isfile(dst):  # Skip cached files
                 paths.append((src_uri, dst))
 
-    # Write new xml file
-    ET.ElementTree(root).write(vrt_destination_path)
+    # Define the output vrt
+    vrt_destination_path = Path(cache_dir, basename(vrt_uri))
 
+    # Copy the new files
     logger.info(f"Downloading {len(paths)} tiles to {cache_dir}")
     for src, dst in paths:
         _copy_to_local(src, dst, fs)
+
+    # Check for the newly cached files
+    new = [item[1] for item in paths]
+    if len(new) == 0:
+        if vrt_destination_path.is_file():
+            return vrt_destination_path
+        else:
+            return Path(vrt_uri)  # The original and try to work with that
+
+    # Second check for existing vrt and add the files to list to build the new vrt
+    if vrt_destination_path.is_file():
+        with open(vrt_destination_path, "r") as f:
+            root = ET.fromstring(f.read())
+        cur = [
+            Path(vrt_destination_path.parent, item.find("SourceFilename").text)
+            for item in root.find("VRTRasterBand").findall(source_name)
+        ]
+        # This shouldnt be possible, but just to be sure.
+        cur = [item for item in cur if item not in new]
+        # Add the current files in the vrt to the list and unlink the current vrt
+        new = new + cur
+        os.unlink(vrt_destination_path)
+
+    # Build the vrt with gdal
+    out_ds = gdal.BuildVRT(
+        destName=vrt_destination_path.as_posix(),
+        srcDSOrSrcDSTab=new,
+    )
+
+    # Close and dereference the gdal dataset
+    out_ds.Close()
+    out_ds = None
+
     return vrt_destination_path